Allow use of an existing connection.
If the `Conn` field is non-nil, operate under the assumption that
the caller has populated it with an already existing connection.

Note: This means some connection fields may have no effect, or not
the fully intended effect.  Specifically, `Timeout` will already be
determined by the existing connection, and `Transport` will only
affect retry in the case of TCP and connection EOF errors.